Refactor Code In The Model, View, And Controller
Added FriendListChanged and SelectedFriendChanged events to Model
to ensure that the Controller / View are always up to date with the
current FriendList and SelectedFriend.

Removed Model.SetSearchMonth() as it was only used once internally
in Model

Refactored Model.FindFriend() to sanitize it's 'name' argument
before performing the search

Refactored View.DisplayFriendList() to display a pre-typed message
when there are no friends to display

Refactored View.DisplaySelectedFriendData() to accept a null value
for it's argument and to clear all data TextBoxes when null is passed

Removed View.ClearSelecetedFriendData() as it has been replaced with
a call to View.DisplaySelectedFriendData(null) with a null argument
as shown

Added a KeyPress listener to the Find TextBox to perform a
friend search when the return key is pressed.

Refactored Find Button Click handler to perform the following tasks:

- Sanitize user input

- Display message to the user when the input is empty, or when
no friend was found